Fix time_t usage for 32bit systems with 64bit time
On newer 32bit systems (e.g. Debian Trixie on arm32)
time_t is a 64bit signed type, so %ld cannot be used
for printf/scanf.
To stay compatible to older systems with 32bit time_t,
always cast time_t to intmax_t for usage with
printf/scanf in the same way as in vdr itself.

This patch fixes a SEGV crash on affected systems.

Unfortunately gcc 14.2 does not check the vasprintf()
format properly (in contrast to asprintf() format).
To verify that all time_t variables are converted
properly, the original epgsearchtools function
msprintf() was replaced with following macro:
 msprintf(strp, args...) \
  ({int res = asprintf(strp, args); \
    if (res < 0) *strp = NULL; \
    res; \
   })
This macro was used for verification only and is not
part of this patch.

+++ epgsearchsvdrp.c
@@ -479,16 +479,16 @@ cString cPluginEpgsearch::SVDRPCommand(const char *Command, const char *Option, 
                 std::string title = pEvent->Title() ? encodeSpecialCharacters(pEvent->Title()) : "";
                 std::string shorttext = pEvent->ShortText() ? encodeSpecialCharacters(pEvent->ShortText()) : "";
 
-                cString cmdbuf = cString::sprintf("%d:%u:%s:%s:%ld:%ld:%s:%ld:%ld:%s:%d",
+                cString cmdbuf = cString::sprintf("%d:%u:%s:%s:%jd:%jd:%s:%jd:%jd:%s:%d",
                                                   result->search->ID,
                                                   pEvent->EventID(),
                                                   title.c_str(),
                                                   shorttext.c_str(),
-                                                  pEvent->StartTime(),
-                                                  pEvent->EndTime(),
+                                                  (intmax_t)(pEvent->StartTime()),
+                                                  (intmax_t)(pEvent->EndTime()),
                                                   CHANNELSTRING(channel),
-                                                  timerMode > 0 ? start : -1,
-                                                  timerMode > 0 ? stop : -1,
+                                                  timerMode > 0 ? (intmax_t)start : -1,
+                                                  timerMode > 0 ? (intmax_t)stop : -1,
                                                   timerMode > 0 ? result->search->BuildFile(pEvent) : "",
                                                   timerMode);
